+++ joe/queue.h
@@ -143,7 +143,7 @@ struct stditem {
 
 /* Allocate an item */
 
-void *alitem(void *list, int itemsize);
+void *alitem(void *list, ptrdiff_t itemsize);
 
 /* Free an item */
